+++ shared/types.ts
@@ -60,4 +60,4 @@ export interface ServiceWithHistory extends ServiceConfig {
   history: HistoryFile[];
 }
 
-export type ServiceMode = 'enabled' | 'unavailable' | 'disabled';
+export type ServiceMode = 'enabled' | 'alwaysok' | 'unavailable' | 'disabled';